requirements.txt
#absl-py==0.9.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==8.0.0
colorama==0.4.3
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
dlib==19.21.1
#en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
gast==0.3.3
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==1.7.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
labelme2coco==0.1.2
lxml==4.6.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
murmurhash==1.0.2
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
networkx==2.5.1
nltk==3.6.2
notebook==6.1.4
numpy==1.19.1
object-detection-0.1
opencv-python==4.3.0.36
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
pixellib==0.6.1
plac==1.1.3
preshed==3.0.2
progressbar==2.5
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.12.2
#pycocotools @ git+https://github.com/philferriere/cocoapi.git@2929bd2ef6b451054755dfd7ceb09278f935f7ad#subdirectory=PythonAPI
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
regex==2021.4.4
requests==2.24.0
scikit-image==0.18.1
scikit-learn==0.23.1
scipy==1.5.2
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
sklearn==0.0
spacy==2.3.2
srsly==1.0.2
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
tifffile==2021.4.8
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
urllib3==1.25.10
wasabi==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
xmljson==0.2.1
yfinance==0.1.59
zipp==3.1.0
